requirements.txt
numpy==1.19.4
Pillow==8.0.1
pycrypto==2.6.1
pycryptodome==3.9.9
